# add array lookup to decoder for notieable perf boost with minimal code change