--- a/src/lib.rs
+++ b/src/lib.rs
@@ -178,10 +178,12 @@ const fn val(c: u8, idx: usize) -> Result<u8, FromHexError> {
         let mut i = 0;
 
         while i < arr.len() {
-            arr[i] = match i as u8 {
-                c @ b'A'..=b'F' => c - b'A' + 10,
-                c @ b'a'..=b'f' => c - b'a' + 10,
-                c @ b'0'..=b'9' => c - b'0',
+            // will not overflow as i < 256 or loop breaks
+            let c = i as u8;
+            arr[i] = match c {
+                b'A'..=b'F' => c - b'A' + 10,
+                b'a'..=b'f' => c - b'a' + 10,
+                b'0'..=b'9' => c - b'0',
                 _ => 255,
             };
             i += 1;
@@ -192,7 +194,7 @@ const fn val(c: u8, idx: usize) -> Result<u8, FromHexError> {
 
     let val = LOOKUP[c as usize];
 
-    if val == 0xff {
+    if val == 255 {
         Err(FromHexError::InvalidHexCharacter {
             c: c as char,
             index: idx,